Address review feedback.
